+++ src/heavy_light_decomposition.hpp
@@ -3,7 +3,7 @@
 #include <map>
 #include <optional>
 #include <vector>
-using i64 = long long;
+using Int = long long;
 
 // Heavy-Light Decomposition
 struct HLDecomp {
@@ -143,10 +143,10 @@ struct HLDecomp {
 };
 
 // Returns the cumulative sum array of edge costs (indexed by Ord).
-std::vector<i64> edge_cost_cumsum(
-    const HLDecomp &hld, const std::map<std::pair<int, int>, i64> &edge_cost) {
+std::vector<Int> edge_cost_cumsum(
+    const HLDecomp &hld, const std::map<std::pair<int, int>, Int> &edge_cost) {
   const int n = hld.n;
-  std::vector<i64> cost(n), acc(n + 1);  // indexed by Ord.
+  std::vector<Int> cost(n), acc(n + 1);  // indexed by Ord.
   // Iterate over the NodeID space.
   for (int v = 0; v < n; ++v) {
     // Calculate the cost from v to parent[v].
@@ -154,7 +154,7 @@ std::vector<i64> edge_cost_cumsum(
     if (p == -1) continue;
     auto it = edge_cost.find(std::pair{std::min(v, p), std::max(v, p)});
     if (it == edge_cost.end()) continue;
-    const i64 cost_to_parent = it->second;
+    const Int cost_to_parent = it->second;
     const int ord = hld.node_to_ord[v];
     cost[ord] = cost_to_parent;
   }
